♻️ refactor: reorganiza páginas de autenticação em componentes e hooks
