+++ src/components/VerificationLoader.tsx
@@ -1,0 +1,17 @@
+// components/VerificationLoader.tsx
+import React from 'react';
+
+interface VerificationLoaderProps {
+  message?: string;
+}
+
+export const VerificationLoader: React.FC<VerificationLoaderProps> = ({
+  message = 'Verificando seu e-mail...',
+}) => (
+  <main className="flex items-center justify-center min-h-screen bg-background">
+    <div className="bg-white p-6 rounded-xl shadow-md text-center">
+      <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-primary mx-auto mb-4"></div>
+      <p className="text-gray-600">{message}</p>
+    </div>
+  </main>
+);
