+++ src/pages/EmailVerification.tsx
@@ -1,45 +1,12 @@
-import { useEffect, useRef } from 'react';
-import { useNavigate, useSearchParams } from 'react-router-dom';
-import api from '../services/api';
-import { toast } from 'react-toastify';
+// EmailVerification.tsx - VERSÃO FINAL CORRIGIDA
+import React from 'react';
+import { useEmailVerification } from '../hooks/useEmailVerification';
+import { VerificationLoader } from '../components/VerificationLoader';
 
-const EmailVerification = () => {
-  const [params] = useSearchParams();
-  const navigate = useNavigate();
-  const hasVerified = useRef(false);
+const EmailVerification: React.FC = () => {
+  useEmailVerification(); // ✅ Limpo e simples
 
-  useEffect(() => {
-    const verifyEmail = async () => {
-      if (hasVerified.current) return;
-      hasVerified.current = true;
-
-      const token = params.get('token');
-      if (!token) {
-        toast.error('Token inválido.');
-        navigate('/login');
-        return;
-      }
-
-      try {
-        await api.post('/auth/verify-email', { token });
-        toast.success('E-mail verificado com sucesso!');
-        navigate('/login');
-      } catch {
-        toast.error('Link expirado ou inválido.');
-        navigate('/verify-pending');
-      }
-    };
-
-    verifyEmail();
-  }, [params, navigate]);
-
-  return (
-    <main className="flex items-center justify-center min-h-screen bg-background">
-      <div className="bg-white p-6 rounded-xl shadow-md">
-        <p>Verificando seu e-mail...</p>
-      </div>
-    </main>
-  );
+  return <VerificationLoader />;
 };
 
 export default EmailVerification;
