+++ src/pages/Login.tsx
@@ -1,154 +1,19 @@
-import React, { useEffect, useState } from 'react';
-import { Link, useNavigate, useLocation } from 'react-router-dom';
-import Typography from '../components/ui/Typography';
-import SubmitButton from '../components/ui/SubmitButton';
-import api from '../services/api';
-import { toast } from 'react-toastify';
-import useAuthStore from '../stores/authStore';
-import PasswordInput from '../components/ui/PasswordInput';
-import GoogleLoginButton from '../components/ui/GoogleLoginButton';
+// Login.tsx
+import React from 'react';
+import { LoginForm } from '../components/LoginForm';
+import { useLogin } from '../hooks/useLogin';
 
 const Login: React.FC = () => {
-  const [formData, setFormData] = useState({ emailOrPhone: '', password: '' });
-  const navigate = useNavigate();
-  const location = useLocation();
-  const from = location.state?.from?.pathname || '/feed';
-
-  const setToken = useAuthStore((state) => state.setToken);
-  const validateToken = useAuthStore((state) => state.validateToken);
-  const user = useAuthStore((state) => state.user);
-  const token = useAuthStore((state) => state.token);
-
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  // ✅ Redireciona só quando o usuário estiver disponível
-  useEffect(() => {
-    if (token && user) {
-      navigate(from, { replace: true });
-    }
-  }, [token, user, from, navigate]);
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setFormData({ ...formData, [e.target.id]: e.target.value });
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (isSubmitting) return;
-
-    setIsSubmitting(true);
-
-    try {
-      const response = await api.post('/auth/session', formData);
-      setToken(response.data.token);
-      localStorage.setItem('token', response.data.token);
-      await validateToken();
-      toast.success('Login realizado com sucesso!');
-      localStorage.removeItem('pendingEmail');
-    } catch (err: any) {
-      if (err.response?.data?.error) {
-        const message = err.response.data.error;
-
-        // Caso o backend retorne mensagem indicando que o e-mail não foi verificado
-        if (message.toLowerCase().includes('não verificado')) {
-          toast.warning('Verifique seu e-mail antes de entrar.');
-          const email = formData.emailOrPhone.includes('@')
-            ? formData.emailOrPhone
-            : null;
-          if (email) localStorage.setItem('pendingEmail', email);
-          navigate('/verify-pending');
-          return;
-        }
-
-        toast.error(message);
-      } else if (err.request) {
-        toast.error('Erro de conexão com o servidor.');
-      } else {
-        toast.error('Erro inesperado ao fazer login.');
-      }
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
+  const { formData, isSubmitting, handleChange, handleSubmit } = useLogin();
 
   return (
     <main className="min-h-screen flex items-center justify-center bg-background px-4 py-12">
-      <div className="w-full max-w-md bg-white p-8 rounded-2xl shadow-lg">
-        {/* Título */}
-        <Typography variant="h2" className="text-primary text-center mb-6">
-          Bem-vindo de volta
-        </Typography>
-
-        {/* Formulário de Login */}
-        <form className="space-y-4" onSubmit={handleSubmit}>
-          <div>
-            <label
-              htmlFor="emailOrPhone"
-              className="block mb-1 text-sm font-medium text-gray-700"
-            >
-              E-mail ou Telefone
-            </label>
-            <input
-              type="text"
-              id="emailOrPhone"
-              value={formData.emailOrPhone}
-              onChange={handleChange}
-              className="w-full px-4 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-primary"
-              placeholder="exemplo@email.com ou (99) 99999-9999"
-              required
-            />
-          </div>
-
-          <div>
-            <PasswordInput
-              type="password"
-              id="password"
-              label="Senha"
-              value={formData.password}
-              onChange={handleChange}
-              placeholder="••••••••"
-              required
-            />
-          </div>
-
-          {/* Link para "Esqueci minha senha" */}
-          <div className="text-right">
-            <Link
-              to="/forgot-password"
-              className="text-sm text-primary font-medium hover:underline"
-            >
-              Esqueci minha senha
-            </Link>
-          </div>
-
-          {/* Botão de Entrar */}
-          <SubmitButton loading={isSubmitting}>Entrar</SubmitButton>
-        </form>
-
-        {/* Divisor */}
-        <div className="flex items-center my-6">
-          <hr className="flex-grow border-t border-gray-300" />
-          <span className="mx-4 text-gray-500 text-sm">Ou se preferir</span>
-          <hr className="flex-grow border-t border-gray-300" />
-        </div>
-
-        <div className="flex flex-col items-center justify-center max-w-md">
-          {/* Botão de login com Google */}
-          <GoogleLoginButton />
-        </div>
-
-        {/* Link para cadastro */}
-        <p className="text-center text-sm text-gray-600 mt-6">
-          Ainda não tem uma conta?{' '}
-          <Link
-            to="/register"
-            className="text-primary font-medium hover:underline"
-          >
-            Criar conta
-          </Link>
-        </p>
-      </div>
+      <LoginForm
+        formData={formData}
+        isSubmitting={isSubmitting}
+        onChange={handleChange}
+        onSubmit={handleSubmit}
+      />
     </main>
   );
 };
